del logs
