+++ miles_plugins/models/hf_attention.py
@@ -4,7 +4,6 @@ from abc import ABC, abstractmethod
 
 import torch
 import torch.distributed as dist
-import torch.nn as nn
 from megatron.core import mpu, tensor_parallel
 from megatron.core.inference.contexts import BaseInferenceContext
 from megatron.core.packed_seq_params import PackedSeqParams
@@ -177,29 +176,3 @@ class HuggingfaceAttention(MegatronModule, ABC):
     @abstractmethod
     def hf_forward(self, hidden_states, packed_seq_params):
         """Huggingface forward function"""
-
-
-def setup_hybrid_cp(model: nn.Module, cp_group: dist.ProcessGroup, cp_rank: int, cp_world_size: int) -> None:
-    """Configure GatedDeltaNet modules for native fla CP instead of all-gather duplication.
-
-    Walks the model tree looking for HuggingfaceAttention submodules that have a
-    ``linear_attn`` child (i.e. DeltaNet layers). For each one it sets the CP
-    metadata so that ``_build_cp_context`` produces a valid context, and flips
-    ``use_native_cp`` so the parent skips the all-gather path.
-    """
-    import logging
-
-    logger = logging.getLogger(__name__)
-    count = 0
-    for module in model.modules():
-        if isinstance(module, HuggingfaceAttention):
-            linear_attn = getattr(module, "linear_attn", None)
-            if linear_attn is not None:
-                linear_attn.cp_group = cp_group
-                linear_attn.cp_rank = cp_rank
-                linear_attn.cp_world_size = cp_world_size
-                module.use_native_cp = True
-                count += 1
-
-    if count > 0:
-        logger.info(f"Configured hybrid CP on {count} DeltaNet modules (fla native state passing)")
